LocateWeatherApp/requirements.txt
Pillow
configparser
sv-ttk
customtkinter
geopy
